Rework JPQLNextExpressions
